requirements.txt
argparse~=1.4.0
numpy~=1.21.2
pandas~=1.3.0
Pillow==8.4.0
scikit_image~=0.18.2
matplotlib~=3.4.2
torch~=1.10.0
torchaudio~=0.10.0
torchvision~=0.11.1
scikit-learn~=1.0.1
tqdm~=4.62.0